@@ -1,5 +1,7 @@
 # -*- coding: utf-8 -*-
 class SnippetsController < ApplicationController
+  before_action :set_snippet, only: %w(show destroy)
+
   # GET /snippets
   # GET /snippets.json
   # GET /snippets.atom
@@ -7,7 +9,7 @@ class SnippetsController < ApplicationController
     @snippets = Snippet.published.date_desc.decorate
 
     respond_to do |format|
-      format.html # index.html.erb
+      format.html # index.html.slim
       format.json { render json: @snippets }
       format.atom { render atom: @snippets }
     end
@@ -16,15 +18,10 @@ class SnippetsController < ApplicationController
   # GET /snippets/1
   # GET /snippets/1.json
   def show
-    snippet = Snippet.find_by_hash_id(params[:id])
-    if snippet.blank?
-      snippet = Snippet.find_by_id(params[:id])
-      return render status: :not_found, text: "404 not found" unless snippet.try(:published?)
-    end
-    @snippet = snippet.decorate
+    @snippet = @snippet.decorate
 
     respond_to do |format|
-      format.html # show.html.erb
+      format.html # show.html.slim
       format.json { render json: @snippet }
     end
   end
@@ -35,7 +32,7 @@ class SnippetsController < ApplicationController
     @snippet = Snippet.new.decorate
 
     respond_to do |format|
-      format.html # new.html.erb
+      format.html # new.html.slim
       format.json { render json: @snippet }
     end
   end
@@ -44,8 +41,8 @@ class SnippetsController < ApplicationController
   def create
     return render_access_denied if current_user.blank?
     content = params[:content].strip
-    snippet = current_user.snippets.build(params[:snippet])
-    snippet.published = params[:commit] == "Public post!"
+    snippet = current_user.snippets.build(snippet_params)
+    snippet.published = params[:commit] == 'public'
 
     content.each_line do |raw_content|
       snippet.messages.build(raw_content: raw_content.chomp)
@@ -65,13 +62,25 @@ class SnippetsController < ApplicationController
   # DELETE /snippets/1
   # DELETE /snippets/1.json
   def destroy
-    if snippet = current_user.snippets.find_by_id(params[:id])
-      snippet.destroy
-    end
+    @snippet.destroy
 
     respond_to do |format|
       format.html { redirect_to root_path, notice: 'Snippet was successfully deleted.' }
       format.json { head :no_content }
     end
+  end
+
+  private
+  def set_snippet
+    @snippet = Snippet.find_by_hash_id(params[:id])
+
+    if @snippet.blank?
+      @snippet = Snippet.find_by_id(params[:id])
+      return render_not_found unless @snippet.try(:published?)
+    end
+  end
+
+  def snippet_params
+    params.require(:snippet).permit(:title)
   end
 end
