@@ -2,30 +2,44 @@ class MessageDecorator < Draper::Decorator
   IMAGE_RE = %r!
     \.(jpe?g|gif|png|bmp|tiff?)\Z
     |
-    \Ahttps://secure.gravatar.com/avatar/
+    \Ahttps:\/\/secure.gravatar.com\/avatar\/
   !ix
 
   delegate_all
+  decorates_association :stars
 
-  # Define presentation-specific methods here. Helpers are accessed through
-  # `helpers` (aka `h`). You can override attributes, for example:
-  #
-  #   def created_at
-  #     helpers.content_tag :span, class: 'time' do
-  #       source.created_at.strftime("%a %m/%d/%y")
-  #     end
-  #   end
-
-  # @todo: link to images when secret
   def content
-    if snippet.published?
-      Rinku.auto_link(h.html_escape(model.content), :all) {|link|
-        link =~ IMAGE_RE ? %(#{link}<br><img src="#{link}" alt="">) : link
-      }
-    else
-      model.content.gsub URI.regexp(%w[http https ftp mailto]) do |uri|
-        h.link_to uri, h.transition_path(to: uri)
+    links = []
+    escaped = h.html_escape(model.content)
+    marked =
+      if snippet.published?
+        Rinku.auto_link(escaped, :all) {|link|
+          links << link if link =~ IMAGE_RE
+          link
+        }
+      else
+        escaped.gsub(URI.regexp(%w[http https ftp mailto])) {|link|
+          links << link if link =~ IMAGE_RE
+          h.link_to(link, h.transition_path(to: link))
+        }
+      end
+    if links.present?
+      marked << '<br>'
+      links.each do |link|
+        uri = snippet.published? ? link : h.transition_path(to: link)
+        src = snippet.published? ? link : h.image_proxy_path(of: link)
+        marked << h.link_to(h.image_tag(src, alt: ''), uri)
       end
     end
+    marked
+  end
+
+  def star_count_tag
+    h.content_tag(:span, star_count, class: 'count')
+  end
+
+  def starred_by_tag
+    starred_by = stars.map(&:image_tag).join
+    h.content_tag(:span, h.raw(starred_by), class: 'starred-by')
   end
 end
