@@ -4,6 +4,8 @@ Iiirc::Application.configure do
   # Code is not reloaded between requests
   config.cache_classes = true
 
+  config.eager_load = true
+
   # Full error reports are disabled and caching is turned on
   config.consider_all_requests_local       = false
   config.action_controller.perform_caching = true
@@ -61,7 +63,9 @@ Iiirc::Application.configure do
   # Send deprecation notices to registered listeners
   config.active_support.deprecation = :notify
 
-  # Log the query plan for queries taking more than this (works
-  # with SQLite, MySQL, and PostgreSQL)
-  # config.active_record.auto_explain_threshold_in_seconds = 0.5
+  # Custom settings
+  config.middleware.use ExceptionNotifier,
+    email_prefix: "[iiirc Exception] ",
+    sender_address: %{"Exception Notifier" <no-reply@iiirc.org>},
+    exception_recipients: %w(ameutau@gmail.com KitaitiMakoto@gmail.com)
 end
