@@ -1,5 +1,5 @@
 Fabricator(:star) do
   user    { Fabricate(:user) }
   message { Fabricate(:message) }
-  count   { 0 }
+  count   { 1 }
 end
