@@ -5,7 +5,7 @@ describe "Snippets" do
   subject { page }
 
   let(:user)    { Fabricate(:user) }
-  let(:snippet) {
+  let!(:snippet) {
     Fabricate(:snippet, user: user) do
       before_validation do |snippet|
         snippet.messages << Fabricate(:message)
@@ -19,6 +19,31 @@ describe "Snippets" do
       expect(page.status_code).to be == 200
     end
 
+    context 'no snippet exists' do
+      it 'show no snippet' do
+        visit snippets_path
+        expect(page).not_to have_css("[id~='snippet_']")
+      end
+    end
+
+    context 'some snippets exist' do
+      it 'show snippets' do
+        visit snippets_path
+        expect(page).to have_css("#snippet_#{snippet.id}")
+      end
+
+      context 'unpublished snippets exist' do
+        before do
+          snippet.update_column :published, false
+        end
+
+        it "don't show unpublished snippets" do
+          visit snippets_path
+          expect(page).not_to have_css("#snippet_#{snippet.id}")
+        end
+      end
+    end
+
     context "when atom feed requested" do
       it "should render atom feed when specified atom with header" do
         page.driver.header 'Accept', 'application/atom+xml'
@@ -29,6 +54,13 @@ describe "Snippets" do
       it "should render atom feed when specified atom with extension" do
         visit snippets_path(format: :atom)
         expect(page.response_headers['Content-Type']).to start_with 'application/atom+xml'
+      end
+
+      it "should render links to hubs of PubSubHubbub" do
+        visit snippets_path(format: :atom)
+        feed = RSS::Parser.parse(page.source)
+        links = feed.links.select {|link| link.rel == 'hub'}
+        expect(links.length).to eq Settings.pubsubhubbub.length
       end
 
       shared_examples "valid atom feed" do
@@ -64,10 +96,46 @@ describe "Snippets" do
     end
   end
 
-  describe "GET /snippet/1" do
-    it "show the requested snippet as @snippet" do
-      visit snippet_path(snippet.id)
-      expect(page).to have_content snippet.title
+  describe 'GET /snippets/1' do
+    subject { page }
+
+    context 'when specified snippet not exist' do
+      it 'should not render' do
+        visit snippet_path id: 0
+        expect(page.status_code).to be(404)
+      end
+    end
+
+    context 'when specified snippet exsits' do
+      it "show the requested snippet as @snippet" do
+        visit snippet_path(snippet.id)
+        expect(page).to have_content snippet.title
+      end
+
+      context 'when specified snippet unpublished' do
+        before do
+          snippet.update_column :published, false
+        end
+
+        it 'show nothing' do
+          visit snippet_path(snippet.id)
+          expect(page.status_code).to be(404)
+        end
+      end
+
+      context 'when specified snippet starred' do
+        let!(:star) { Fabricate(:star, message: snippet.messages.first) }
+
+        it 'show user who starred' do
+          visit snippet_path(snippet)
+          expect(page).to have_css(".starred-by img[data-user-id='#{star.user.id}']")
+        end
+
+        it 'icon of who starred should be clickable' do
+          visit snippet_path(snippet)
+          expect(page).to have_css(".starred-by a[href='#{user_path(star.user.username)}'] img")
+        end
+      end
     end
 
     it "respond with 404 when snippet not exist" do
@@ -93,8 +161,6 @@ describe "Snippets" do
     end
 
     context "when current_user is present" do
-      let(:user) { Fabricate(:user) }
-
       before do
         sign_in(user)
         visit new_snippet_path
@@ -119,10 +185,10 @@ describe "Snippets" do
         expect(find('h2')).to have_content 'sugoi'
 
         expect(find('time')).to have_content '00:52'
-        expect(find('span.nick_0')).to have_content 'shikakun:'
+        expect(find("span.nick_3")).to have_content 'shikakun:'
         expect(find('p#L1.tweet')).to have_content 'すごい！'
 
-        expect(current_path).to eq snippet_path(id: 1)
+        expect(current_path).to eq snippet_path(id: snippet.id + 1)
       end
 
       it "should create a snippet without title" do
@@ -135,10 +201,10 @@ describe "Snippets" do
         expect(find('h2')).to have_content(/\A\d{4}-\d{2}-\d{2}\s\d{2}:\d{2}:\d{2}.*\z/)
 
         expect(find('time')).to have_content '00:52'
-        expect(find('span.nick_0')).to have_content 'shikakun:'
+        expect(find("span.nick_3")).to have_content 'shikakun:'
         expect(find('p#L1.tweet')).to have_content 'すごい！'
 
-        expect(current_path).to eq snippet_path(id: 1)
+        expect(current_path).to eq snippet_path(id: snippet.id + 1)
       end
 
       it "should create a snippet as private" do
@@ -151,7 +217,7 @@ describe "Snippets" do
         expect(find('h2')).to have_content(/\A\d{4}-\d{2}-\d{2}\s\d{2}:\d{2}:\d{2}.*\z/)
 
         expect(find('time')).to have_content('00:52')
-        expect(find('span.nick_0')).to have_content('shikakun:')
+        expect(find("span.nick_3")).to have_content 'shikakun:'
         expect(find('p#L1.tweet')).to have_content('すごい！')
 
         expect(current_path).to match(/\A\/snippets\/\w{20}\z/)
@@ -168,14 +234,35 @@ describe "Snippets" do
     end
 
     context "when user is logged in" do
-      before { sign_in(user) }
-
-      it "should show delete button and delete snippet" do
+      before do
+        sign_in(user)
         visit snippet_path(snippet.id)
-        expect(page).to have_content "Delete Snippet!"
-        click_on "Delete Snippet!"
-        expect(page).to have_content "Snippet was successfully deleted."
-        expect(current_path).to be == root_path
+      end
+
+      shared_examples 'delete button' do
+        it 'should show delete button and delete snippet' do
+          expect(page).to have_content "Delete Snippet!"
+          click_on "Delete Snippet!"
+          expect(page).to have_content "Snippet was successfully deleted."
+          expect(current_path).to be == root_path
+        end
+      end
+
+      it_behaves_like 'delete button'
+
+      context "when snippet is secret" do
+        let!(:snippet) {
+          Fabricate(:snippet, user: user, published: false) do
+            before_validation do |snippet|
+              snippet.messages << Fabricate(:message)
+            end
+          end
+        }
+        before do
+          visit snippet_path(snippet.hash_id)
+        end
+
+        it_behaves_like 'delete button'
       end
     end
   end
